clinisys/requirements.txt
# Core dependencies
pandas>=1.3.0
numpy>=1.21.0
duckdb>=0.9.2
pyyaml>=5.4.1
sqlalchemy>=1.4.0

# For logging, os, sys, datetime, glob, hashlib, collections, re, typing: all are standard library

# If using Jupyter/IPython for notebooks:
jupyter
ipython 